requirements.txt
# Git interaction
gitpython>=3.1.0

# Time zone handling
pytz>=2024.1

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning
torch>=2.0.0
transformers>=4.36.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0 